Fix inpgrid inpgrid call in config and reorganise some components
