+++ docs/source/swan/components/group.rst
@@ -10,6 +10,7 @@ SWAN group commands
 
    rompy.swan.components.group.BaseGroupComponent
    rompy.swan.components.group.STARTUP
+   rompy.swan.components.group.INPGRIDS
    rompy.swan.components.group.PHYSICS
    rompy.swan.components.group.OUTPUT
    rompy.swan.components.group.LOCKUP
