+++ docs/source/swan/components/inpgrid.rst
@@ -9,7 +9,6 @@ SWAN input grids commands
    :nosignatures:
    :toctree: _generated/
 
-   rompy.swan.components.inpgrid.INPGRIDS
    rompy.swan.components.inpgrid.REGULAR
    rompy.swan.components.inpgrid.CURVILINEAR
    rompy.swan.components.inpgrid.UNSTRUCTURED
