+++ docs/source/swan/index.rst
@@ -27,16 +27,6 @@ Data
     rompy.swan.boundary.DataBoundary
 
 
-Forcing
--------
-
-.. autosummary::
-    :nosignatures:
-    :toctree: _generated/
-
-    rompy.swan.forcing.ForcingData
-
-
 Components
 ----------
 
@@ -48,20 +38,21 @@ models called `subcomponents` to handle more complex arguments.
 Components are subclasses of :py:class:`rompy.swan.components.base.BaseComponent`.
 The base component class implements the following attribues:
 
-* The **model_type** field that must be overwritten in each component subclass. The
-  `model_type` field is defined as a `Literal`_ type and is used to specify the exact
-  component class in a declarative framework (i.e., when using a dict from a yaml or
-  json file to prescribe the model config).
+* The **model_type** field that should be overwritten in each component subclass. The
+  `model_type` field is defined as a `Literal`_ type and is used to discriminate the
+  exact components to use in fields defined by a `Union` type of two or more components
+  in a declarative framework (i.e., instantiating with a dict from yaml or json file).
 
 * The **cmd()** method that must be overwritten in each component subclass. The `cmd()`
   method should return either a string or a list of strings to fully define a SWAN
   command line instruction. A list of strings defines multiple command line
-  instructions that are executed in sequence such as the [INPGRID, READGRID] components.
+  instructions that are executed in sequence such as the INPGRID/READGRID components.
 
 * The **render()** method that constructs the command line instruction from the content
   returned from the `cmd()` method. The `render()` method is typically called inside
-  the model template to construct the specific command line instruction from that
-  component, taking care of maximum line size, line break and line continuation.
+  the `__call__` method of the config class to construct the specific command line
+  instruction from that component, taking care of maximum line size, line break and
+  line continuation.
 
 
 Components are defined within the :py:mod:`rompy.swan.components` subpackage and
@@ -122,6 +113,24 @@ arguments to one or more component. The following modules are available:
 * :doc:`subcomponents/physics`
 * :doc:`subcomponents/numerics`
 * :doc:`subcomponents/output`
+
+
+Interface
+---------
+
+Interface classes provide an interface between swan components and higher level objects
+such as `TimeRange`, `Data` and `Grid` objects. They are used inside the `__call__`
+method of the config classes to pass instances of these objects to the appropriate
+components and define consistent parameters to the config after instantiating them.
+
+
+.. autosummary::
+    :nosignatures:
+    :toctree: _generated/
+
+    rompy.swan.interface.ForcingData
+    rompy.swan.interface.OutputInterface
+    rompy.swan.interface.LockupInterface
 
 
 Types
